requirements.txt
flask
torch
torchvision
Pillow
requests
ultralytics
pandas
